Add files via upload
This Python program creates a basic calculator with the ability to add, subtract, multiply, and divide. It asks the user to choose an operation and two numbers, then evaluates the operation and displays the output. The program also allows the user to perform multiple calculations in a row.

Source: Conversation with Bing, 11/28/2023
(1) Python Program to Make a Simple Calculator. https://www.programiz.com/python-programming/examples/calculator.
(2) Simple Calculator Program in Python - W3Schools. https://www.w3schools.in/python/examples/simple-calculator.
(3) How To Make a Calculator Program in Python 3 | DigitalOcean. https://www.digitalocean.com/community/tutorials/how-to-make-a-calculator-program-in-python-3.